评论

+++ src/main/java/com/wan/communtify/controller/QuestionController.java
@@ -15,7 +15,7 @@ public class QuestionController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("question/{id}")
-    public String question(@PathVariable("id") Integer id, Model model){
+    public String question(@PathVariable("id") Long id, Model model){
         QuestionDTO questionDTO=questionService.getById(id);
         //累加阅读数
         questionService.incView(id);
